Remove redundant 'static_cast' (#4437)
* Modify the annotation: fix the label serial number mistake


* Remove redundant 'static_cast'
Dimension construct funtion need int64_t type
And tsp.dim(i).dim_value() is int64_t type
Do not need static_cast<int>

+++ onnx/common/ir_pb_converter.cc
@@ -190,7 +190,7 @@ std::vector<Dimension> tensorShapeProtoToDimensions(const ONNX_NAMESPACE::Tensor
   dims.reserve(tsp.dim_size());
   for (int i = 0; i < tsp.dim_size(); i++) {
     if (tsp.dim(i).has_dim_value()) {
-      dims.emplace_back(static_cast<int>(tsp.dim(i).dim_value()));
+      dims.emplace_back(tsp.dim(i).dim_value());
     } else if (tsp.dim(i).has_dim_param()) {
       dims.emplace_back(tsp.dim(i).dim_param());
     } else {
